第五轮修复战役：加载场景对象不显示双根因修复 + 第四轮遗留清零（双架构构建通过）
面板/视口双根因（用户报告"加载新场景后对象不正确显示"）：
- pmm_load_v1 逐模型标志用 x86 裸偏移写入，x64 上 loadComplete 永不置位，
  UpdateModelVertexBuffers 首行 gate 早退 → PMD 蒙皮在场景加载后从未运行
  （T-pose/网格爆炸/toon 错乱混合分布）；全站类型化，AllStar 9/9 与原版
  对照 posed 正确（双进程骨骼 dump 证明骨骼层本就逐位一致）
- ui_panel_sweep/paint/editor_click 的面板行簿记 x86 裸偏移：-999 清零扫
  写坏 x64 boneCount/facialFrameCount → 选中模型场景加载必崩；model_layout
  新增 boneListSelLine/RowType/RowRecord 命名字段（双 ABI static_assert 钉位
  x86 11708/11712/11912、x64 0x3158/0x315C/0x3224）

第四轮遗留清零：
- PMX 蒙皮完整移植：附加 UV 0-4 五 stride worker、SDEF（点积符号修正 nlerp
  非 slerp）、BDEF4 双累加序、PMX 顶点 morph 恢复/施加段补齐；PMD 路径零改动
- Kinect：阶段字段脑裂统一（泵块 m_kinectCaptureStage ↔ state.autoRepeat），
  补泵序言块（深度图请求 + 菜单 0x124 使能卫生）；物理门 A 占位改共享
  state.messageSeen 计数器（0xA54CC/0xA1B98 定谳为已移植功能，占位系误判）
- effect 帧渲染序列对齐 x64 0x1400C1E60：帧头 FILLMODE/LIGHTING 移至背景/
  地面之后、模型走查前矩阵重设 + ALPHABLEND、帧尾 DESTBLEND 收尾对
- fpsLimit 布局法证：尾部真实漂移 1112 字节（pad 误尺寸累积，含三处整块
  误诊纠正），97 个 x64 pin，sizeof==0xA55E0 双架构精确断言；顺带修
  comboSelIndex 裸读（x64 读错字段）
- 单精度化八站：贝塞尔早退/slerp/IK CCD/物理回写链/WinMain/追赶/colBytes/
  滚轮，均指令级对齐；slerp 钳制常量勘误 0.99999994f→0.999999f；追赶补
  物理模式 count 门
- MMDxShow x64 版：布局双架构断言（VC8 x86/VC10 x64 两列），CMake 门控
  解除，29 项 COM 冒烟 port-x86/x64 与原版两版四方一致
- effect_api.cpp 13 处裸偏移类型化（x64 effect 导出 API 曾全线读错字段）；
  刚体组行步长 0x2E 修复（步长 1 漏检）

工具沉淀：tools/mdump2.py + mdiff.py（双进程骨骼层 dump/diff 回归工具）

+++ src/window/ui_scroll_mouse.cpp
@@ -133,9 +133,12 @@ void HandleVScroll(LPARAM lParam, WPARAM wParam) {
 // The original re-fetches GetDlgItem(.., 427) before every call (the
 // handle is identical each time).
 //
-// Otherwise the wheel scales a float by delta * 0.05 - the multiply is
-// done in double against the 0.05f constant (bit-exact with dbl_52D738),
-// the result is stored back to float (fmul/fadd/fstp dword):
+// Otherwise the wheel scales a float by delta * 0.05 - x86 runs
+// fild/fmul dbl_52D738/fadd/fstp dword (double multiply, one rounding at
+// the float store); the x64 twin sub_7FF7CB45F300 is pure SSE single
+// precision (movsx/movd/cvtdq2ps, mulss against the dword slot
+// 0x7FF7CB552CB0 whose image value IS 0x3D4CCCCD = 0.05f, addss the
+// field, movss back) - the port follows the x64 form:
 //   camera mode (this+760 set AND this+656432 >= 0) -> this+828 (0x33C)
 //   morph-follow special branch (this+760 == 0 && this+650648 &&
 //       this+816 == 0 && this+656432 == this+2320 && this+656432 >= 0):
@@ -172,10 +175,10 @@ void HandleMouseWheel(int delta) {
     const std::uint8_t mode = app->state.optflag[0];  // 760
 
     if ((mode & (axis >= 0 ? 1u : 0u)) != 0) {
-        // camera distance: this+828 (float) += delta * 0.05
-        app->CameraPosition()[2] = static_cast<float>(
-            static_cast<double>(wheel) * 0.05f +
-            static_cast<double>(app->CameraPosition()[2]));
+        // camera distance: this+828 (float) += wheel * 0.05
+        // (x64 0x7FF7CB45F340: cvtdq2ps/mulss 0.05f/addss/movss)
+        app->CameraPosition()[2] =
+            static_cast<float>(wheel) * 0.05f + app->CameraPosition()[2];
     } else {
         const bool morphFollow =
             mode == 0 &&
@@ -192,10 +195,10 @@ void HandleMouseWheel(int delta) {
             PostViewRefresh(app);                                 // 0x40D130
             return;
         }
-        // view angle: this+657628 (float) += delta * 0.05
-        app->CameraDistance() = static_cast<float>(
-            static_cast<double>(wheel) * 0.05f +
-            static_cast<double>(app->CameraDistance()));
+        // view angle: this+657628 (float) += wheel * 0.05
+        // (x64 0x7FF7CB45F410: same single-precision chain)
+        app->CameraDistance() =
+            static_cast<float>(wheel) * 0.05f + app->CameraDistance();
     }
     RefreshRequest(-1);       // 0x440AC0
     PostViewRefresh(app);     // 0x40D130
